fix: remove component itemsbreadcrumb
