--- src/stories/components/Breadcrumb/ItemsBreadcrumb/types.ts
@@ -1,8 +1,0 @@
-export interface Items {
-  text: string;
-  redirectLink?: string;
-}
-
-export interface ItemsBreadcrumbProps {
-  items: Items[];
-}
