+++ src/stories/components/Breadcrumb/index.tsx
@@ -1,17 +1,24 @@
-import React, { FC, useMemo } from 'react';
+import React, { FC } from 'react';
 
-import { BreadcrumbsProps } from './types';
 import BreadcrumbsMui from '@mui/material/Breadcrumbs';
-import { ItemsBreadcrumb } from './ItemsBreadcrumb';
+
+import * as S from './styles';
+import { BreadcrumbsProps } from './types';
 
 export const Breadcrumbs: FC<BreadcrumbsProps> = ({ items, separator = '>' }) => {
-  const breadcrumbs = useMemo(() => {
-    return ItemsBreadcrumb({ items });
-  }, [items]);
-
   return (
     <BreadcrumbsMui separator={separator} aria-label="breadcrumb">
-      {breadcrumbs}
+      {items?.map((item, index) => {
+        const isLastItem = index + 1 === items.length;
+        if (isLastItem && item.redirectLink === undefined) {
+          return <S.LastText key={item.text}>{item.text}</S.LastText>;
+        }
+        return (
+          <S.LinkLabel key={item.text} href={item.redirectLink ?? ''}>
+            {item.text}
+          </S.LinkLabel>
+        );
+      })}
     </BreadcrumbsMui>
   );
 };
